+++ src/context.mli
@@ -86,25 +86,24 @@ type t =
   ; (** Native dynlink *)
     natdynlink_supported : bool
 
-  ; ocamlc_config           : Ocamlc_config.t
+  ; ocaml_config            : Ocaml_config.t
   ; version_string          : string
   ; version                 : int * int * int
   ; stdlib_dir              : Path.t
   ; ccomp_type              : string
   ; c_compiler              : string
-  ; ocamlc_cflags           : string
-  ; ocamlopt_cflags         : string
-  ; bytecomp_c_libraries    : string
-  ; native_c_libraries      : string
-  ; native_pack_linker      : string
-  ; ranlib                  : string
-  ; cc_profile              : string
+  ; ocamlc_cflags           : string list
+  ; ocamlopt_cflags         : string list
+  ; bytecomp_c_libraries    : string list
+  ; native_c_libraries      : string list
+  ; cc_profile              : string list
   ; architecture            : string
   ; system                  : string
   ; ext_obj                 : string
   ; ext_asm                 : string
   ; ext_lib                 : string
   ; ext_dll                 : string
+  ; ext_exe                 : string
   ; os_type                 : string
   ; default_executable_name : string
   ; host                    : string
@@ -129,12 +128,8 @@ val sexp_of_t : t -> Sexp.t
 (** Compare the context names *)
 val compare : t -> t -> Ordering.t
 
-(** If [use_findlib] is [false], don't try to guess the library search path with opam or
-    ocamlfind. This is only for building jbuilder itself, so that its build is completely
-    independent of the user setup. *)
 val create
-  :  ?use_findlib:bool
-  -> ?merlin:bool
+  :  ?merlin:bool
   -> Workspace.Context.t
   -> t list Fiber.t