+++ src/main.mli
@@ -6,12 +6,12 @@ type setup =
   ; (* Evaluated jbuilds per context names *)
     stanzas      : (Path.t * Scope_info.t * Stanzas.t) list String_map.t
   ; contexts     : Context.t list
-  ; packages     : Package.t String_map.t
+  ; packages     : Package.t Package.Name.Map.t
   ; file_tree    : File_tree.t
   }
 
 (* Returns [Error ()] if [pkg] is unknown *)
-val package_install_file : setup -> string -> (Path.t, unit) result
+val package_install_file : setup -> Package.Name.t -> (Path.t, unit) result
 
 (** Scan the source tree and discover everything that's needed in order to build
     it. *)
@@ -20,14 +20,14 @@ val setup
   -> ?filter_out_optional_stanzas_with_missing_deps:bool
   -> ?workspace:Workspace.t
   -> ?workspace_file:string
-  -> ?only_packages:String_set.t
+  -> ?only_packages:Package.Name.Set.t
   -> ?x:string
   -> ?ignore_promoted_rules:bool
   -> unit
   -> setup Fiber.t
 val external_lib_deps
   : ?log:Log.t
-  -> packages:string list
+  -> packages:Package.Name.t list
   -> unit
   -> Build.lib_deps Path.Map.t