+++ src/ocamldep.ml
@@ -6,39 +6,41 @@ module SC = Super_context
 module Dep_graph = struct
   type t =
     { dir        : Path.t
-    ; per_module : (unit, Module.t list) Build.t String_map.t
+    ; per_module : (unit, Module.t list) Build.t Module.Name.Map.t
     }
 
   let deps_of t (m : Module.t) =
-    match String_map.find t.per_module m.name with
+    match Module.Name.Map.find t.per_module m.name with
     | Some x -> x
     | None ->
       Sexp.code_error "Ocamldep.Dep_graph.deps_of"
         [ "dir", Path.sexp_of_t t.dir
-        ; "modules", Sexp.To_sexp.(list string) (String_map.keys t.per_module)
-        ; "module", Sexp.atom m.name
+        ; "modules", Sexp.To_sexp.(list Module.Name.t)
+                       (Module.Name.Map.keys t.per_module)
+        ; "module", Module.Name.t m.name
         ]
 
   module Dep_closure =
-    Top_closure.Make(String)(struct
+    Top_closure.Make(Module.Name)(struct
       type t = Module.t
-      type graph = t list String_map.t
+      type graph = t list Module.Name.Map.t
       let key (t : t) = t.name
-      let deps t map = Option.value_exn (String_map.find map (key t))
+      let deps t map = Option.value_exn (Module.Name.Map.find map (key t))
     end)
 
   let top_closed t modules =
     Build.all
-      (List.map (String_map.to_list t.per_module) ~f:(fun (unit, deps) ->
+      (List.map (Module.Name.Map.to_list t.per_module) ~f:(fun (unit, deps) ->
          deps >>^ fun deps -> (unit, deps)))
     >>^ fun per_module ->
-    let per_module = String_map.of_list_exn per_module in
+    let per_module = Module.Name.Map.of_list_exn per_module in
     match Dep_closure.top_closure per_module modules with
     | Ok modules -> modules
     | Error cycle ->
-      die "dependency cycle between modules in %s:\n   %s" (Path.to_string t.dir)
-        (String.concat ~sep:"\n-> "
-           (List.map cycle ~f:Module.name))
+      die "dependency cycle between modules in %s:\n   %a"
+        (Path.to_string t.dir)
+        (Fmt.list ~pp_sep:Fmt.nl (Fmt.prefix (Fmt.string "-> ") Module.Name.pp))
+        (List.map cycle ~f:Module.name)
 
   let top_closed_implementations t modules =
     Build.memoize "top sorted implementations" (
@@ -48,7 +50,7 @@ module Dep_graph = struct
 
   let dummy (m : Module.t) =
     { dir = Path.root
-    ; per_module = String_map.singleton m.name (Build.return [])
+    ; per_module = Module.Name.Map.singleton m.name (Build.return [])
     }
 end
 
@@ -83,10 +85,11 @@ let parse_deps ~dir ~file ~(unit : Module.t)
         String.extract_blank_separated_words
           (String.sub line ~pos:(i + 1) ~len:(String.length line - (i + 1)))
         |> List.filter_map ~f:(fun m ->
+          let m = Module.Name.of_string m in
           if m = unit.name then
             None
           else
-            String_map.find modules m)
+            Module.Name.Map.find modules m)
       in
       (match lib_interface_module with
        | None -> ()
@@ -98,15 +101,17 @@ let parse_deps ~dir ~file ~(unit : Module.t)
          in
          if unit.name <> m.name && not is_alias_module &&
             List.exists deps ~f:(fun x -> Module.name x = m.name) then
-           die "Module %s in directory %s depends on %s.\n\
+           die "Module %a in directory %s depends on %a.\n\
                 This doesn't make sense to me.\n\
                 \n\
-                %s is the main module of the library and is \
+                %a is the main module of the library and is \
                 the only module exposed \n\
                 outside of the library. Consequently, it should \
                 be the one depending \n\
                 on all the other modules in the library."
-             unit.name (Path.to_string dir) m.name m.name);
+             Module.Name.pp unit.name (Path.to_string dir)
+             Module.Name.pp m.name
+             Module.Name.pp m.name);
       let deps =
         match alias_module with
         | None -> deps
@@ -115,16 +120,16 @@ let parse_deps ~dir ~file ~(unit : Module.t)
       deps
 
 let rules ~(ml_kind:Ml_kind.t) ~dir ~modules
-      ?(already_used=String_set.empty)
+      ?(already_used=Module.Name.Set.empty)
       ~alias_module ~lib_interface_module sctx =
   let per_module =
-    String_map.map modules ~f:(fun unit ->
+    Module.Name.Map.map modules ~f:(fun unit ->
       match Module.file ~dir unit ml_kind with
       | None -> Build.return []
       | Some file ->
         let ocamldep_output = Path.extend_basename file ~suffix:".d" in
         let context = SC.context sctx in
-        if not (String_set.mem already_used unit.name) then
+        if not (Module.Name.Set.mem already_used unit.name) then
           SC.add_rule sctx
             (Build.run ~context (Ok context.ocamldep)
                [A "-modules"; Ml_kind.flag ml_kind; Dep file]
@@ -137,7 +142,7 @@ let rules ~(ml_kind:Ml_kind.t) ~dir ~modules
   let per_module =
     match alias_module with
     | None -> per_module
-    | Some m -> String_map.add per_module m.name (Build.return [])
+    | Some m -> Module.Name.Map.add per_module m.name (Build.return [])
   in
   { Dep_graph.
     dir