+++ test/blackbox-tests/test-cases/copy_files/run.t
@@ -1,17 +1,17 @@
   $ $JBUILDER build -j1 test.exe .merlin --display short --root . --debug-dependency-path
       ocamllex lexers/lexer1.ml
-      ocamldep test.ml.d
       ocamldep lexer1.ml.d
+      ocamldep test.ml.d
       ocamldep dummy.ml.d
-        ocamlc bar.o
         ocamlc .foo.objs/dummy.{cmi,cmo,cmt}
-    ocamlmklib dllfoo_stubs.so,libfoo_stubs.a
         ocamlc .test.eobjs/lexer1.{cmi,cmo,cmt}
-      ocamlopt .foo.objs/dummy.{cmx,o}
         ocamlc .test.eobjs/test.{cmi,cmo,cmt}
+        ocamlc bar.o
+    ocamlmklib dllfoo_stubs.so,libfoo_stubs.a
+      ocamlopt .foo.objs/dummy.{cmx,o}
       ocamlopt .test.eobjs/lexer1.{cmx,o}
-      ocamlopt foo.{a,cmxa}
       ocamlopt .test.eobjs/test.{cmx,o}
+      ocamlopt foo.{a,cmxa}
       ocamlopt test.exe
   $ $JBUILDER build -j1 @bar-source --display short --root .
   #line 1 "include/bar.h"