+++ CHANGES.md
@@ -5,6 +5,29 @@ next
 
 - Reduce the number of simultaneously opened fds (#578)
 
+- Add a workaround for when a library normally installed by the
+  compiler is not installed but still has a META file (#574, fixes
+  #563)
+
+- Do not depend on ocamlfind. Instead, hard-code the library path when
+  installing from opam (#575)
+
+- Reduce interleaving in the scheduler in an attempt to make Jbuilder
+  keep file descriptors open for less long (#586)
+
+- Accept and ignore `ppx.driver` fields in library stanzas, in
+  preparation for the generic ppx driver system (#588)
+
+- Change the default behavior regarding the check for overlaps between
+  local and installed libraries. Now even if there is no link time
+  conflict, we don't allow an external dependency to overlap with a
+  local library, unless the user specifies `allow_overlapping_dependencies`
+  in the jbuild file (#587, fixes #562)
+
+- Expose a few more variables in jbuild files: `ext_obj`, `ext_asm`,
+  `ext_lib`, `ext_dll` and `ext_exe` as well as `${ocaml-config:XXX}`
+  for most variables in the output of `ocamlc -config` (#590)
+
 1.0+beta18 (25/02/2018)
 -----------------------